API

Types

PropDicts.PropDict

Functions and macros

PropDicts.trim_null!

Documentation

PropDicts.PropDict — Type

PropDict <: AbstractDict{Union{Symbol,Int},Any}

PropDicts is a dictionary for that supports

Constructors:

PropDict(dict::AbstractDict)

PropDict(key1 => value, key2 => value2, ...)

During construction, keys are automatically converted to Symbols and Ints, values that are dicts are converted to PropDicts.

PropDict support deep merging:

x = PropDict(:a => PropDict(:b => 7, :c => 5, :d => 2), :e => "foo")
y = PropDict(:a => PropDict(:c => 42, :d => nothing), :f => "bar")

z = merge(x, y)
@assert z == PropDict(
    :a => PropDict(:b=>7, :d => nothing, :c => 42),
    :e => "foo", :f => "bar"
)

PropDicts.trim_null!(z)
@assert z == PropDict(
    :a => PropDict(:b=>7, :c => 42),
    :e => "foo", :f => "bar"
)

PropDicts.trim_null! — Function

trim_null!(d::AbstractDict; recursive::Bool = true)

Remove values equal to nothing from d.

Operates recursively on values in d if recursive == true.